Cygwin: read ud_t struct members instead of calling accessors.
This should result in smaller and faster code, since the accessors are
not inline.  It also allows using the 'pc' field, which is the pc at the
end of the current instruction (as opposed to ud_insn_off which was the
pc at the start of the instruction), removing the need for the extra
addition of ud_insn_len and therefore the inline helper rip_rel_offset
function.

Also, use a define for pointer size in bits.  The proliferation of magic
'64' values was kind of ugly.  There are still a number of magic '32'
values, but these are the size of the immediate or displacement of the
instruction, and thus determine which branch of the lval union to use.

Signed-off-by: Jeremy Drake <cygwin@jdrake.com>